[INLONG-10096][Manager] Add doc
